ieee802154: add valid frame length helper
This patch adds a generic valid frame length check function helper. This
is useful to check the length field after receiving. For example the
at86rf231 doesn't filter invalid frame length. Sometimes the CRC can also
be correct. If we get the lqi value with an invalid frame length the
kernel may crash, because we dereference an invalid pointer.